FPM-00 Add implementation at SecurityConfiguration
