+++ src/main/java/com/ecomerceApi/Priscila/model/User.java
@@ -37,11 +37,6 @@ public class User implements UserDetails {
     }
 
     @Override
-    public String getPassword() {
-        return password;
-    }
-
-    @Override
     public String getUsername() {
         return email;
     }
