+++ src/main/java/com/ecomerceApi/Priscila/repository/UserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> { // Long because our Id is type long.
+public interface UserRepository extends JpaRepository<User, Long> { // Long because our id is type long.
 
     Optional<User> findByEmail(String email);
 }
