+++ src/main/java/com/ecomerceApi/Priscila/requestModels/LoginRequest.java
@@ -2,10 +2,11 @@ package com.ecomerceApi.Priscila.requestModels;
 
 import lombok.Data;
 import lombok.Getter;
+import lombok.Value;
 
-@Data
-@Getter
+@Value
 public class LoginRequest {
-    private String username;
-    private String password;
+
+    String username;
+    String password;
 }
