+++ src/main/java/com/ecomerceApi/Priscila/requestModels/RegisterRequest.java
@@ -11,6 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class RegisterRequest {
+
     private String name;
     private String email;
     private String password;
